document-sync/src/egrid/DelphiEgrid/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
atlassian-python-api
pandas
langchain
python-dotenv
PyPDF2
azure-core
azure-storage-blob
azure-search
azure-search-documents==11.4.0b3
azure-ai-formrecognizer
azure-cosmos
tiktoken
nltk
openpyxl
